refactored burger constructor
