+++ .TODO
@@ -1,8 +1,4 @@
-* Если нажать кнопку заказа, а потом выбрать булку - оформится заказ
-* Добавить подсказки при оформлении заказа, если не выбрана булка или ингридиенты
-* Исправить баг с переключением табов - при переключении крайних табов
 * Переписать проект на Redux Toolkit
-* Мемоизировать весь проект
 * Сделать возможность перетаскивать ингредиенты из левого блока в правый с возможностью сортировки сразу
 * Все пустые строки в стейтах поменять на null
 * Типизировать setCookie.ts
@@ -18,11 +14,4 @@
 * Проверить оформляется ли заказ с протухшим токеном
 * Типизировать requestCreator
 * После оформления заказа и перезагрузки страницы модалка заказа открыта. Исправить
-* Создать компонент для булок
-* Сделать вложенность папок components
-
-Какие компоненты нужно мемоизировать и переработать:
-* BurgerConstructor
-* BurgerIngredients
-* BurgerItem
-* ConstructorItem
+* fetch('https://norma.nomoreparties.space/api/orders')
