+++ public/index.html
@@ -17,7 +17,6 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <div id="modal"></div>
     <div id="popup-hint"></div>
   </body>
 </html>
