+++ src/components/BurgerConstructor/BurgerConstructor.module.scss
@@ -3,6 +3,8 @@
 @import 'scss/vars';
 
 .wrapper {
+	height: 100%;
+
 	display: flex;
 	flex-direction: column;
 	justify-content: space-between;
@@ -10,45 +12,4 @@
 	row-gap: 40px;
 
 	grid-area: cart;
-
-	height: 100%;
-}
-
-
-
-
-
-
-
-.choiceOffer {
-	width: 100%;
-	height: 100%;
-	display: flex;
-	flex-direction: row;
-	justify-content: center;
-	align-items: center;
-}
-
-.orderBox {
-	display: flex;
-	flex-direction: row;
-	justify-content: flex-end;
-	align-items: center;
-	column-gap: 40px;
-
-	padding: 0 16px;
-}
-
-.price {
-	@include font-digits-medium;
-
-	display: flex;
-	flex-direction: row;
-	align-items: center;
-	column-gap: 8px;
-
-	svg {
-		width: 33px;
-		height: 33px;
-	}
 }
