+++ src/components/BurgerConstructor/ConstructorBody/ConstructorBody.module.scss
@@ -20,16 +20,6 @@
 	}
 }
 
-.bun {
-	width: 100%;
-	padding: 0 16px;
-
-	display: flex;
-	flex-direction: row;
-	align-items: center;
-	justify-content: flex-end;
-}
-
 .ingredients {
 	width: 100%;
