+++ src/components/BurgerConstructor/ConstructorBody/ConstructorBody.tsx
@@ -1,15 +1,13 @@
 import React from 'react'
-import { useSelector } from 'react-redux'
 import { useDrop } from 'react-dnd'
-import { ConstructorElement } from '@ya.praktikum/react-developer-burger-ui-components'
 
 // Redux
-import { useDispatch } from 'redux/store'
-import { cartSelector, dataIngreientsSelector } from 'redux/selectors'
+import { useDispatch, useSelector } from 'redux/store'
 import { setBun, setIngredient } from 'redux/actionCreators'
+import { cartSelector, dataIngreientsSelector } from 'redux/selectors'
 
 // Components
-import { ConstructorItem, ConstructorPlug } from 'components'
+import { ConstructorBun, ConstructorItem, ConstructorPlug } from 'components'
 
 // Styles
 import styles from './ConstructorBody.module.scss'
@@ -18,7 +16,7 @@ const ConstructorBody: React.FC = () => {
   const dispatch = useDispatch()
 
   const allIngredients = useSelector(dataIngreientsSelector)
-  const { bun, ingredients } = useSelector(cartSelector)
+  const { ingredients } = useSelector(cartSelector)
 
   // Burger content
   const burgerIngredients = React.useMemo(
@@ -59,19 +57,7 @@ const ConstructorBody: React.FC = () => {
 				${isHover ? styles.wrapper_hover : ''}
 			`}
     >
-      <div className={styles.bun}>
-        {bun ? (
-          <ConstructorElement
-            type="top"
-            isLocked={true}
-            text={`${bun.name} (верх)`}
-            price={bun.price}
-            thumbnail={bun.image}
-          />
-        ) : (
-          <ConstructorPlug position="top" title="Выберите булку" />
-        )}
-      </div>
+      <ConstructorBun position="top" />
 
       <div className={styles.ingredients}>
         {burgerIngredients.length ? (
@@ -83,19 +69,7 @@ const ConstructorBody: React.FC = () => {
         )}
       </div>
 
-      <div className={styles.bun}>
-        {bun ? (
-          <ConstructorElement
-            type="bottom"
-            isLocked={true}
-            text={`${bun.name} (низ)`}
-            price={bun.price}
-            thumbnail={bun.image}
-          />
-        ) : (
-          <ConstructorPlug position="bottom" title="Выберите булку" />
-        )}
-      </div>
+      <ConstructorBun position="bottom" />
     </div>
   )
 }
