+++ src/components/BurgerConstructor/ConstructorBun/ConstructorBun.module.scss
@@ -1,0 +1,11 @@
+//? === Constructor Bun ===
+
+.bun {
+	width: 100%;
+	padding: 0 16px;
+
+	display: flex;
+	flex-direction: row;
+	align-items: center;
+	justify-content: flex-end;
+}
