+++ src/components/index.ts
@@ -21,3 +21,5 @@ export { default as AuthLink } from './AuthLink/AuthLink'
 export { default as ConstructorBody } from './BurgerConstructor/ConstructorBody/ConstructorBody'
 export { default as Router } from './Router/Router'
 export { default as Portal } from './Portal/Portal'
+export { default as ConstructorFooter } from './BurgerConstructor/ConstructorFooter/ConstructorFooter'
+export { default as ConstructorBun } from './BurgerConstructor/ConstructorBun/ConstructorBun'
